定期更新

+++ src/Application.php
@@ -14,14 +14,23 @@ namespace think\swoole;
 use Swoole\Http\Request;
 use Swoole\Http\Response;
 use think\App;
+use think\Db;
 use think\Error;
 use think\exception\HttpException;
+use think\facade\Config;
+use think\facade\Env;
 
 /**
  * Swoole应用对象
  */
 class Application extends App
 {
+    // 请求控制 chan
+    private $chan = null;
+
+    // Websocket请求控制 chan
+    private $wsChan = null;
+
     /**
      * 处理Swoole请求
      * @access public
@@ -31,12 +40,52 @@ class Application extends App
      */
     public function swoole(Request $request, Response $response)
     {
+        if (empty($this->chan)) {
+            $this->chan = new \chan(1);
+        }
+        $this->chan->push(1);
         try {
+
+            $users=Db::name('user')->select();
+            echo "go start\n";
+            $chan = new \chan(1);
+            \go(function ()use($chan) {
+                echo "co before select user \n";
+                Db::name('user')->select();
+                echo "co after select user \n";
+//                $chan->push(1);
+            });
+
+            echo "go end\n";
+
+//            $chan->pop();
+
+
+            // 测试用代码
+//            $date = date('Y-m-d H:i:s');
+//            $this->chan->push($date);
+//            echo $date . "\Swoole\Coroutine::getuid():" . \Swoole\Coroutine::getuid() . "\n";
+//            \Swoole\Coroutine::sleep(10);
+//            echo $date . "this->chan->pop;\n";
+//            $date = $this->chan->pop();
+            $uri = $request->server['request_uri'];
+            if ($uri == '/favicon.ico') {
+                $response->status(404);
+                $response->end();
+                return;
+            }
+
             ob_start();
 
             // 重置应用的开始时间和内存占用
             $this->beginTime = microtime(true);
             $this->beginMem  = memory_get_usage();
+
+            // 重置数据库查询次数
+            Db::$queryTimes = 0;
+
+            // 重置数据库执行次数
+            Db::$executeTimes = 0;
 
             // 销毁当前请求对象实例
             $this->delete('think\Request');
@@ -49,12 +98,21 @@ class Application extends App
             $_POST   = $request->post ?: [];
             $_FILES  = $request->files ?: [];
             $header  = $request->header ?: [];
+            $server  = $request->server ?: [];
+
             if (isset($header['x-requested-with'])) {
-                $header['http_x_requested_with'] = $header['x-requested-with'];
+                $server['HTTP_X_REQUESTED_WITH'] = $header['x-requested-with'];
             }
 
-            $server  = array_change_key_case($request->server, CASE_UPPER);
-            $_SERVER = array_merge($server, array_change_key_case($header, CASE_UPPER));
+            if (isset($header['referer'])) {
+                $server['http_referer'] = $header['referer'];
+            }
+
+            if (isset($_GET[$this->config->get('var_pathinfo')])) {
+                $server['path_info'] = $_GET[$this->config->get('var_pathinfo')];
+            }
+
+            $_SERVER = array_change_key_case($server, CASE_UPPER);
 
             // 重新实例化请求对象 处理swoole请求数据
             $this->request->withHeader($header)
@@ -104,7 +162,10 @@ class Application extends App
                 $response->header($key, $val);
             }
 
-            $response->end($content);
+            $response->write($content);
+            $response->end();
+
+
         } catch (HttpException $e) {
             $this->exception($response, $e);
         } catch (\Exception $e) {
@@ -112,10 +173,16 @@ class Application extends App
         } catch (\Throwable $e) {
             $this->exception($response, $e);
         }
+
+        $this->chan->pop();
     }
 
     public function swooleWebSocket($server, $frame)
     {
+        if (empty($this->wsChan)) {
+            $this->wsChan = new \chan(1);
+        }
+        $this->wsChan->push(1);
         try {
             // 重置应用的开始时间和内存占用
             $this->beginTime = microtime(true);
@@ -126,14 +193,17 @@ class Application extends App
             WebSocketFrame::destroy();
             $request = $frame->data;
             $request = json_decode($request, true);
+
             // 重置应用的开始时间和内存占用
             $this->beginTime = microtime(true);
             $this->beginMem  = memory_get_usage();
             WebSocketFrame::getInstance($server, $frame);
-            $_COOKIE                    = isset($request['arguments']['cookie']) ? $request['arguments']['cookie'] : [];
-            $_GET                       = isset($request['arguments']['get']) ? $request['arguments']['get'] : [];
-            $_POST                      = isset($request['arguments']['post']) ? $request['arguments']['post'] : [];
-            $_FILES                     = isset($request['arguments']['files']) ? $request['arguments']['files'] : [];
+
+            $_COOKIE = isset($request['arguments']['cookie']) ? $request['arguments']['cookie'] : [];
+            $_GET    = isset($request['arguments']['get']) ? $request['arguments']['get'] : [];
+            $_POST   = isset($request['arguments']['post']) ? $request['arguments']['post'] : [];
+            $_FILES  = isset($request['arguments']['files']) ? $request['arguments']['files'] : [];
+
             $_SERVER["PATH_INFO"]       = $request['url'] ?: '/';
             $_SERVER["REQUEST_URI"]     = $request['url'] ?: '/';
             $_SERVER["SERVER_PROTOCOL"] = 'http';
@@ -145,12 +215,11 @@ class Application extends App
                 ->withGet($_GET)
                 ->withPost($_POST)
                 ->withCookie($_COOKIE)
-                ->withInput($request->rawContent())
                 ->withFiles($_FILES)
-                ->setBaseUrl($request->server['request_uri'])
-                ->setUrl($request->server['request_uri'] . (!empty($request->server['query_string']) ? '&' . $request->server['query_string'] : ''))
-                ->setHost($request->header['host'])
-                ->setPathinfo(ltrim($request->server['path_info'], '/'));
+                ->setBaseUrl($request['url'])
+                ->setUrl($request['url'])
+                ->setHost(Config::get("app_host"))
+                ->setPathinfo(ltrim($request['url'], '/'));
 
             // 更新请求对象实例
             $this->route->setRequest($this->request);
@@ -165,6 +234,7 @@ class Application extends App
         } catch (\Throwable $e) {
             $this->webSocketException($server, $frame, $e);
         }
+        $this->wsChan->pop();
     }
 
     protected function exception($response, $e)
@@ -190,7 +260,7 @@ class Application extends App
     protected function webSocketException($server, $frame, $e)
     {
         $response = [
-            'code'    => $e->code,
+            'code'    => $e->getCode(),
             'content' => $e->getMessage(),
         ];
         $server->push($frame->fd, json_encode($response));
